requirements.txt
amqp==5.1.1
asgiref==3.6.0
attrs==23.1.0
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cron-descriptor==1.4.0
cryptography==41.0.0
Django==4.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-crontab==0.7.1
django-db==0.0.7
django-phonenumber-field==7.1.0
django-rest-swagger==2.2.0
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.3.0
MarkupSafe==2.1.2
mysqlclient==2.1.1
openapi-codec==1.3.2
packaging==23.1
phonenumbers==8.13.11
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
pyrsistent==0.19.3
python-crontab==2.7.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
requests==2.31.0
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
swagger-spec-validator==3.0.3
typing-extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
zipp==3.15.0
